Merged PR 50: User Authorization and Forgot Password
Implemented the AuthorizeUser attribute filter, which will ensure the Session contains a user object and a token string and that the token belongs to the user. If so, it will allow the Controller method to continue.
I implemented the necessary methods in the AccountController, DbHelper, and created a ResetCodes table in the database so that when a user submits their email in the ForgotPassword flow, it will link a temporary code to their user ID and include this code in the URL within the email. When they follow the link, it will verify the code and allow them to set a new password.